requirements.txt
dataclasses==0.6
Poutyne==0.6
matplotlib==3.1.0
torch==1.3.1
torchvision==0.4.2
numpy==1.16.4
imgaug==0.3.0
torchsummary==1.5.1
Pillow==6.2.1
